file split + penalty commands
