+++ models/weather.py
@@ -1,0 +1,134 @@
+from datetime import datetime, timedelta
+
+from globals import SUNSET_TIME, SUNRISE_TIME
+from utils.common_utils import hours_to_hhmm, convert_to_timezone, format_datetime, seconds_to_verbose_interval, get_rain_eta_irl_time
+class Weather:
+    def __init__(self, name: str, emoji: str, day_thumbnail: str, night_thumbnail: str):
+        self.name = name
+        self.emoji = emoji
+        self.day_thumbnail = day_thumbnail
+        self.night_thumbnail = night_thumbnail
+
+
+# Represents GTA game time and its related attributes
+class GTATime:
+    def __init__(self, hours_game_time: float, weekday, weather_period_time: float):
+        self.hours_game_time = hours_game_time
+        self.str_game_time = hours_to_hhmm(self.hours_game_time)
+        self.weekday = weekday
+        self.weather_period_time = weather_period_time
+        self.is_day_time = SUNRISE_TIME <= self.hours_game_time < SUNSET_TIME
+
+
+# Represents rain estimation details (ETA)
+class RainETA:
+    def __init__(self, sec_eta: int, is_raining: bool):
+        self.sec_eta = sec_eta
+        self.str_eta = seconds_to_verbose_interval()
+        self.is_raining = is_raining
+
+
+    def get_rain_eta_irl_time(self, current_time: datetime, timezone_str: str) -> str:
+        return get_rain_eta_irl_time(self.sec_eta, current_time, timezone_str)
+
+class WeatherState:
+    """Represents the full current weather conditions"""
+    def __init__(self, weather_instance: Weather, gta_time: GTATime, rain_eta: RainETA):
+        self.weather = weather_instance
+        self.gta_time = gta_time
+        self.rain_eta = rain_eta
+
+# Weather states with all conditions
+WEATHER_STATES = {
+    'clear': Weather(
+        "Clear", "☀️", "https://i.imgur.com/LerUU1Z.png", "https://i.imgur.com/waFNkp1.png"
+    ),
+    'raining': Weather(
+        "Raining", "🌧️", "https://i.imgur.com/qsAl41k.png", "https://i.imgur.com/jc98A0G.png"
+    ),
+    'drizzling': Weather(
+        "Drizzling", "🌦️", "https://i.imgur.com/Qx18aHp.png", "https://i.imgur.com/EWSCz5d.png"
+    ),
+    'misty': Weather(
+        "Misty", "🌁", "https://i.imgur.com/mjZwX2A.png", "https://i.imgur.com/Mh1PDXS.png"
+    ),
+    'foggy': Weather(
+        "Foggy", "🌫️", "https://i.imgur.com/mjZwX2A.png", "https://i.imgur.com/Mh1PDXS.png"
+    ),
+    'hazy': Weather(
+        "Hazy", "🌫️", "https://i.imgur.com/mjZwX2A.png", "https://i.imgur.com/Mh1PDXS.png"
+    ),
+    'snowy': Weather(
+        "Snowy", "❄️", "https://i.imgur.com/WJEjWM6.png", "https://i.imgur.com/1TxfthS.png"
+    ),
+    'cloudy': Weather(
+        "Cloudy", "☁️", "https://i.imgur.com/1oMUp2V.png", "https://i.imgur.com/qSOc8XX.png"
+    ),
+    'mostly_cloudy': Weather(
+        "Mostly cloudy", "🌥️", "https://i.imgur.com/aY4EQhE.png", "https://i.imgur.com/2LIbOFC.png"
+    ),
+    'partly_cloudy': Weather(
+        "Partly cloudy", "⛅", "https://i.imgur.com/aY4EQhE.png", "https://i.imgur.com/2LIbOFC.png"
+    ),
+    'mostly_clear': Weather(
+        "Mostly clear", "🌤️", "https://i.imgur.com/aY4EQhE.png", "https://i.imgur.com/2LIbOFC.png"
+    )
+}
+
+# Weather state changes based on the game time
+WEATHER_STATE_CHANGES = [
+    [0, WEATHER_STATES['partly_cloudy']],
+    [4, WEATHER_STATES['misty']],
+    [7, WEATHER_STATES['mostly_cloudy']],
+    [11, WEATHER_STATES['clear']],
+    [14, WEATHER_STATES['misty']],
+    [16, WEATHER_STATES['clear']],
+    [28, WEATHER_STATES['misty']],
+    [31, WEATHER_STATES['clear']],
+    [41, WEATHER_STATES['hazy']],
+    [45, WEATHER_STATES['partly_cloudy']],
+    [52, WEATHER_STATES['misty']],
+    [55, WEATHER_STATES['cloudy']],
+    [62, WEATHER_STATES['foggy']],
+    [66, WEATHER_STATES['cloudy']],
+    [72, WEATHER_STATES['partly_cloudy']],
+    [78, WEATHER_STATES['foggy']],
+    [82, WEATHER_STATES['cloudy']],
+    [92, WEATHER_STATES['mostly_clear']],
+    [104, WEATHER_STATES['partly_cloudy']],
+    [105, WEATHER_STATES['drizzling']],
+    [108, WEATHER_STATES['partly_cloudy']],
+    [125, WEATHER_STATES['misty']],
+    [128, WEATHER_STATES['partly_cloudy']],
+    [131, WEATHER_STATES['raining']],
+    [134, WEATHER_STATES['drizzling']],
+    [137, WEATHER_STATES['cloudy']],
+    [148, WEATHER_STATES['misty']],
+    [151, WEATHER_STATES['mostly_cloudy']],
+    [155, WEATHER_STATES['foggy']],
+    [159, WEATHER_STATES['clear']],
+    [176, WEATHER_STATES['mostly_clear']],
+    [196, WEATHER_STATES['foggy']],
+    [201, WEATHER_STATES['partly_cloudy']],
+    [220, WEATHER_STATES['misty']],
+    [222, WEATHER_STATES['mostly_clear']],
+    [244, WEATHER_STATES['misty']],
+    [247, WEATHER_STATES['raining']],
+    [250, WEATHER_STATES['drizzling']],
+    [252, WEATHER_STATES['partly_cloudy']],
+    [268, WEATHER_STATES['misty']],
+    [270, WEATHER_STATES['partly_cloudy']],
+    [272, WEATHER_STATES['cloudy']],
+    [277, WEATHER_STATES['partly_cloudy']],
+    [292, WEATHER_STATES['misty']],
+    [295, WEATHER_STATES['partly_cloudy']],
+    [300, WEATHER_STATES['mostly_cloudy']],
+    [306, WEATHER_STATES['partly_cloudy']],
+    [318, WEATHER_STATES['mostly_cloudy']],
+    [330, WEATHER_STATES['partly_cloudy']],
+    [337, WEATHER_STATES['clear']],
+    [367, WEATHER_STATES['partly_cloudy']],
+    [369, WEATHER_STATES['raining']],
+    [376, WEATHER_STATES['drizzling']],
+    [377, WEATHER_STATES['partly_cloudy']]
+]
